The interrupt routine is finally working!!! but...
The interrupt routine is working, but not in the most appropriate way. We must ensure that the program continues to run as usual, even after an interrupt. 

Currently, if we press the key during the yellow light routine, the program will correctly execute the green light routine until it ends. However, after that, the program will return to the exact point where the interupt was triggered, which means the yellow light routine will not be fully executed again.